refactor: 폴더 구조 변경으로 인한 파일 위치 조정
